# src/server/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Second Admin API",
    "version": "0.3.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    },
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "PageSize": {
        "name": "pageSize",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
      "CSRFToken": {
        "name": "X-CSRF-Token",
        "in": "header",
        "required": false,
        "description": "Required for unsafe Cookie-authenticated requests.",
        "schema": {
          "type": "string"
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        }
      },
...
      "IDsRequest": {
        "type": "object",
        "required": [
          "ids"
        ],
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
...
      "Menu": {
        "type": "object",
        "required": [
          "id",
          "parentId",
          "type",
          "name",
          "sort",
          "visible",
          "status"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "parentId": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "type": "string",
            "enum": [
              "directory",
              "menu",
              "button"
            ]
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "component": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "sort": {
            "type": "integer"
          },
          "visible": {
            "type": "boolean"
          },
          "permission": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Menu"
            }
          }
        }
      },
      "MenuRequest": {
        "type": "object",
        "required": [
          "parentId",
          "type",
          "name",
          "path",
          "component",
          "icon",
          "sort"
        ],
        "properties": {
          "parentId": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "type": "string",
            "enum": [
              "directory",
              "menu",
              "button"
            ]
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "component": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "sort": {
            "type": "integer"
          },
          "visible": {
            "type": "boolean"
          },
          "permission": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "enum": [
              0,
              1
            ]
          }
        }
      },
      "CurrentMenus": {
        "type": "object",
        "required": [
          "menus",
          "permissions"
        ],
        "properties": {
          "menus": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Menu"
            }
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Dictionary": {
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "status"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          }
        }
      },
      "DictionaryRequest": {
        "type": "object",
        "required": [
          "code",
          "name"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "enum": [
              0,
              1
            ]
          }
        }
      },
      "DictionaryItem": {
        "type": "object",
        "required": [
          "id",
          "dictionaryId",
          "label",
          "value",
          "sort",
          "status"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "dictionaryId": {
            "type": "integer",
            "format": "int64"
          },
          "label": {
            "type": "string"
          },
          "value": {
            "type": "string"
          },
          "sort": {
            "type": "integer"
          },
          "status": {
            "type": "integer"
          }
        }
      },
      "DictionaryItemRequest": {
        "type": "object",
        "required": [
          "label",
          "value",
          "sort"
        ],
        "properties": {
          "label": {
            "type": "string"
          },
          "value": {
            "type": "string"
          },
          "sort": {
            "type": "integer"
          },
          "status": {
            "type": "integer",
            "enum": [
              0,
              1
            ]
          }
        }
      },
      "OperationLog": {
        "type": "object",
        "required": [
          "id",
          "userId",
          "requestId",
          "method",
          "path",
          "statusCode",
          "durationMs",
          "ip",
          "userAgent",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "requestId": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "statusCode": {
            "type": "integer"
          },
          "durationMs": {
            "type": "integer",
            "format": "int64"
          },
          "ip": {
            "type": "string"
          },
          "userAgent": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "OperationLogPage": {
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OperationLog"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "LoginLog": {
        "type": "object",
        "required": [
          "id",
          "username",
          "event",
          "success",
          "ip",
          "userAgent",
          "deviceId",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "event": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "ip": {
            "type": "string"
          },
          "userAgent": {
            "type": "string"
          },
          "deviceId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "LoginLogPage": {
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LoginLog"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      }
    },
    "responses": {
...
      "Conflict": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/menus": {
      "get": {
        "operationId": "listMenus",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Menu tree",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Menu"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createMenu",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CSRFToken"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MenuRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Menu created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Menu"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/menus/current": {
      "get": {
        "operationId": "currentMenus",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current routes and button permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentMenus"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/menus/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64"
          }
        }
      ],
      "put": {
        "operationId": "updateMenu",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CSRFToken"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MenuRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Menu updated"
          }
        }
      },
      "delete": {
        "operationId": "deleteMenu",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CSRFToken"
          }
        ],
        "responses": {
          "204": {
            "description": "Menu deleted"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/api/v1/roles/{id}/menus": {
      "put": {
        "operationId": "setRoleMenus",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "$ref": "#/components/parameters/CSRFToken"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IDsRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Role menus replaced"
          }
        }
      }
    },
    "/api/v1/dictionaries": {
      "get": {
        "operationId": "listDictionaries",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Dictionaries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Dictionary"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createDictionary",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CSRFToken"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictionaryRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Dictionary created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dictionary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dictionaries/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64"
          }
        }
      ],
      "put": {
        "operationId": "updateDictionary",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CSRFToken"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictionaryRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Dictionary updated"
          }
        }
      },
      "delete": {
        "operationId": "deleteDictionary",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CSRFToken"
          }
        ],
        "responses": {
          "204": {
            "description": "Dictionary deleted"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/api/v1/dictionaries/{id}/items": {
      "post": {
        "operationId": "createDictionaryItem",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "$ref": "#/components/parameters/CSRFToken"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictionaryItemRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Dictionary item created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DictionaryItem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dictionaries/{code}/items": {
      "get": {
        "operationId": "dictionaryItems",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Enabled dictionary items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DictionaryItem"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dictionary-items/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64"
          }
        }
      ],
      "put": {
        "operationId": "updateDictionaryItem",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CSRFToken"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictionaryItemRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Dictionary item updated"
          }
        }
      },
      "delete": {
        "operationId": "deleteDictionaryItem",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CSRFToken"
          }
        ],
        "responses": {
          "204": {
            "description": "Dictionary item deleted"
          }
        }
      }
    },
    "/api/v1/operation-logs": {
      "get": {
        "operationId": "listOperationLogs",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated operation logs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationLogPage"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/login-logs": {
      "get": {
        "operationId": "listLoginLogs",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated login logs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginLogPage"
                }
              }
            }
          }
        }
      }
    }
  }
}